% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibMuso.R
\name{calibMuso}
\alias{calibMuso}
\title{calibMuso}
\usage{
calibMuso(settings,parameters=NULL, timee="d", debugging=FALSE, logfilename=NULL,
keepEpc=FALSE, export=FALSE, silent=FALSE, aggressive=FALSE, leapYear=FALSE)
}
\arguments{
\item{settings}{You have to run the setupMuso function before calibMuso. It is its output which contains all of the necessary system variables. It sets the whole running environment}

\item{parameters}{In the settings variable you have set the row indexes of the variables, you wish to change. In this parameter you can give an exact value for them in a vector like: c(1,2,3,4)}

\item{timee}{The required timesteps in the modell output. It can be "d", if it is daily, "m", if it's monthly, "y", it it is yearly. I recommend to use daily data, the yearly and monthly data is not well-tested yet.}

\item{debugging}{If it is TRUE, it copies the log file to a Log directory to store it, if it is stamplog it contatenate a number before the logfile, which is one more than the maximum of the represented ones in the LOG directory. If it is true or stamplog it collects the "wrong" logfiles}

\item{logfilename}{If you want to set a specific name for your logfiles you can set this via logfile parameter}

\item{keepEpc}{If TRUE, it keeps the epc file and stamp it, after these copies it to the EPCS directory. If debugging True or false, it copies the wrong epc files to the wrong epc directory.}

\item{export}{if it is yes or you give a filename here, it converts the ouxtput to the specific extension. For example, if you set export to "example.csv", it converts the output to "csv", if you set it to "example.xls" it converts to example.xls with the xlsx package. If it is not installed it gives back a warning message and converts it to csv.}

\item{silent}{If you set it TRUE all off the modells output to the screen will be suppressed. It can be usefull, because it increases the model-speed.}

\item{aggressive}{It deletes every possible modell-outputs from the previous modell runs.}

\item{keepBinary}{In default RBBGCMuso to keep  working area as clean as possible, deletes all the regular output files. The results are directly printed to the standard output, but you can redirect it, and save it to a variable, or you can export your results to the desired destination in a desired format. Whith this variable you can enable to keep the binary output files. If you want to set the location of the binary output, please take a look at the binaryPlace argument.}

\item{binaryPlace}{The place of the binary output files.}

\item{fileToChange}{You can change any line of the epc, soil or the ini file, you just have to specify with this variable which file you want to change. Three options possible: "epc", "ini", "soil"}

\item{skipSpinup}{If TRUE, calibMuso wont do spinup simulation}

\item{prettyOut}{date ad Date type, separate year, month, day vectors}

\item{leapYearHandling}{Should the function do a leapyear correction on the outputdata? If TRUE, then the dec 31 is removed for every leap year}
}
\value{
No return, outputs are written to file
}
\description{
This function changes the epc file and after that  runs the BBGC-MuSo model and reads in its outputfile in a well-structured way.
}
\author{
Roland Hollos
}
