% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeMuso.R
\name{changemulline}
\alias{changemulline}
\title{This is the function which is capable to change multiple specific lines to others using their row numbers.}
\usage{
changemulline(filePaths, calibrationPar, contents, fileOut, fileToChange,
  modifyOriginal = FALSE)
}
\description{
The function uses the previous changspecline function to operate.
}
\author{
Roland Hollos
}
\keyword{internal}
