% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{changeMusoC}
\alias{changeMusoC}
\title{changeMusoC}
\usage{
changeMusoC(inFile, outFile, inMat)
}
\arguments{
\item{inFile}{is the big matrix}

\item{outFile}{is the small matrix}
}
\description{
This function is fastly randomize values based on min and max values
}
