% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/changcontent.R
\name{changspecline}
\alias{changspecline}
\title{This function calls the UNIX(-like) sed program to change specific line to other, using the row numbers.}
\usage{
changspecline(filename, line_number, content)
}
\arguments{
\item{The}{name of the file which is needed to be changed in some lines, the numbers of this lines(vector), and
the contents(vector).}
}
\value{
void
}
\description{
This function calls the UNIX(-like) sed program to change specific line to other, using the row numbers.
}
\author{
Roland
}

