% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSite.R
\name{compareCalibratedWithOriginal}
\alias{compareCalibratedWithOriginal}
\title{compareCalibratedWithOriginal}
\usage{
compareCalibratedWithOriginal(
  key,
  modOld,
  modNew,
  mes,
  likelihoods,
  alignIndexes,
  musoCodeToIndex,
  nameGroupTable,
  groupFun
)
}
\arguments{
\item{key}{keyword}
}
\description{
This functions compareses the likelihood and the RMSE values of the simulations and the measurements
}
