% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{dayOfMonths}
\alias{dayOfMonths}
\title{dayOfMonths}
\usage{
dayOfMonths(year, corrigated=TRUE)
}
\arguments{
\item{year}{a year}

\item{corrigated}{Do you want to handle the leapyears, if yes choose TRUE}
}
\value{
vector with 12 element. First is January, the last is December. All of the vector element represents the number of the days in that specific month
}
\description{
This function gives as a vector which contains the number of the days per each month
}
\examples{
dayOfMonths(2004, corrigated=TRUE)

dayOfMonths(2004, corrigated=FALSE)

}
