% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{dayOfYears}
\alias{dayOfYears}
\title{This function tells us how many days are in the given year.}
\usage{
dayOfYears(year, corrigated = TRUE)
}
\description{
This function tells us how many days are in the given year.
}
\author{
Roland Hollos
}
\keyword{internal}
