% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUsefullFunctions.R
\name{fextension}
\alias{fextension}
\title{fextension}
\usage{
fextension(filename)
}
\arguments{
\item{filename}{The string of the filenam}
}
\value{
the extension of the given file
}
\description{
A function for extracting the extension name from the filename string
}
\author{
Roland Hollos
}
