% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUsefullFunctions.R
\name{file.path2}
\alias{file.path2}
\title{file.path2}
\usage{
file.path2(str1, str2)
}
\arguments{
\item{str1}{This is the first path string}

\item{str2}{This is the second path string}
}
\value{
A concatenated path
}
\description{
It is an extended file.path function, it can concatenate path where the first ends and the second begins with "/", so
there wont be two  slash nearby eachother
}
\author{
Roland Hollos
}
