% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUsefullFunctions.R
\name{fixAlloc}
\alias{fixAlloc}
\title{fixAlloc}
\usage{
...
}
\arguments{
\item{settings}{the base RMuso settings variable}

\item{type}{normal or spinup depending what you want to modify}
}
\description{
Fix allocation parameter in the epc file
}
