% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.R
\name{flatMuso}
\alias{flatMuso}
\title{flatMuso}
\usage{
flatMuso(
  iniName,
  execPath = "./",
  depTree = options("RMuso_depTree")[[1]],
  directory = "flatdir",
  d = TRUE,
  outE = TRUE
)
}
\arguments{
\item{iniName}{The name of the ini file}

\item{depTree}{The file dependency defining dataframe. At default it is:  options("RMuso_depTree")[[1]]}

\item{directory}{The destination directory for flattening. At default it will be flatdir}
}
\description{
This function reads the ini file and creates a directory (named after the directory argument) with all the files the modell uses with this file. the directory will be flat.
}
