% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genEpc.R
\name{genEpc}
\alias{genEpc}
\title{genEpc}
\usage{
genEpc(sourceEpc, targetEpc, parameters)
}
\arguments{
\item{sourceEpc}{the original epc file-the template}

\item{parameters}{dataframe where in the first column there are the indices and the second column the values}

\item{location}{output location directory}
}
\description{
randEpc is a random epc creator based on musoMonte
}
\author{
Roland HOLLOS
}
