% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUsefullFunctions.R
\name{getConstMatrix}
\alias{getConstMatrix}
\title{getConstMatrix}
\usage{
getConstMatrix(filetype = "epc",
  version = as.character(getOption("RMuso_version")))
}
\arguments{
\item{filetype}{It can be "epc" or "soil".}

\item{version}{The version of the MuSo environment}
}
\description{
getConstMatrix is a function whith wich you can get the default constrain matrix for your choosen type and version.
}
