% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.R
\name{getFilesFromIni}
\alias{getFilesFromIni}
\title{getFilesFromIni}
\usage{
getFilesFromIni(
  iniName,
  execPath = "./",
  depTree = options("RMuso_depTree")[[1]]
)
}
\arguments{
\item{iniName}{The name of the ini file}

\item{depTree}{The file dependency defining dataframe. At default it is:  options("RMuso_depTree")[[1]]}
}
\description{
This function reads the ini file and gives yout back the path of all file involved in model run
}
