% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistantFunctions.R
\name{getLogs}
\alias{getLogs}
\title{getLogs}
\usage{
getLogs(outputLoc, outputNames, type = "spinup")
}
\arguments{
\item{outputLoc}{This is the location of the output files.}

\item{outputNames}{These are the prefixis of the logfiles}
}
\value{
Logfiles with paths
}
\description{
This function gives us the muso logfiles with their paths
}
\keyword{internal}
