% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/other_usefull_functions.R
\name{getyearlymax}
\alias{getyearlymax}
\title{'Function for getting the maximum values of the years, from daily data}
\usage{
getyearlymax(daily_observations)
}
\arguments{
\item{A}{vector of the daily observations}
}
\value{
A vector of yearly data
}
\description{
'Function for getting the maximum values of the years, from daily data
}
\author{
Roland Hollós
}

