% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{isLeapyear}
\alias{isLeapyear}
\title{isLeapyear}
\usage{
isLeapyear(year)
}
\arguments{
\item{year}{a year}
}
\value{
TRUE, if leapyear, FALSE if dont.
}
\description{
This function tells us if its argument a leapyear or not.
}
\examples{
isLeapyear(2004)
isLeapyear(2000)
isLeapyear(2100)
}
