% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metReportMuso.R
\name{metReportMuso}
\alias{metReportMuso}
\title{metReportMuso}
\usage{
metReportMuso(metFileName, skip = 4, numericReport = FALSE)
}
\arguments{
\item{metFileName}{The name of the meteorology file (mtc43).}

\item{skip}{Number of header lines in meteorology file.}

\item{numericReport}{If numericReport is set to FALSE, the function returns with a text report. If numericReport is set to TRUE, the function returns with a numeric report.}
}
\value{
It depends on the numericReport parameter. The function returns with a text report, or with a numeric report.
}
\description{
This function calculates the daily and yearly statistics for a given meteorology file (mtc43).
}
\author{
Erzsebet Kristof
}
