% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSite.R
\name{multiSiteThread}
\alias{multiSiteThread}
\title{multiSiteThread}
\usage{
multiSiteThread(
  measuredData,
  parameters = NULL,
  startDate = NULL,
  endDate = NULL,
  formatString = "\%Y-\%m-\%d",
  calTable,
  dataVar,
  outLoc = "./calib",
  outVars = NULL,
  iterations = 300,
  skipSpinup = TRUE,
  plotName = "calib.jpg",
  modifyOriginal = TRUE,
  likelihood,
  uncertainity = NULL,
  burnin = NULL,
  naVal = NULL,
  postProcString = NULL,
  threadNumber,
  constraints = NULL,
  th = 10,
  fileToModify = NULL,
  fileToChange = "epc"
)
}
\description{
This is an
}
\author{
Roland HOLLOS
}
