% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuso.R
\name{musoCalculateAxisBreaks}
\alias{musoCalculateAxisBreaks}
\title{x axis range helper function}
\usage{
musoCalculateAxisBreaks(
  date_vector,
  year_axis_interval_base = 2,
  silent = TRUE
)
}
\description{
x axis range helper function
}
