% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeMuso.R
\name{musoCompareFiles}
\alias{musoCompareFiles}
\title{musoCompareFiles}
\usage{
musoCompareFiles(filenames, indices)
}
\arguments{
\item{filenames}{The name of the files where you can get the data from}

\item{indices}{muso indices}
}
\description{
A simple wrapper function based on musoGetValues where you can get multiple values from multiple files 
using the supplied indices. It is useful for comparing files.
}
