% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{musoDate}
\alias{musoDate}
\title{musoDate}
\usage{
musoDate(
  startYear,
  endYears = NULL,
  numYears,
  combined = TRUE,
  leapYearHandling = FALSE,
  prettyOut = FALSE
)
}
\description{
This function generates MuSo compatibla dates for the data
}
\author{
Roland HOLLOS
}
