% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeMuso.R
\name{musoGetValues}
\alias{musoGetValues}
\title{musoGetValues}
\usage{
musoGetValues(filename, indices)
}
\arguments{
\item{filename}{The name of the musofile we want the value from (e.g. epc file)}

\item{indices}{muso indices}
}
\description{
Get values from a musofile by supplying muso indices
}
