% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{musoGlue}
\alias{musoGlue}
\title{musoGlue}
\usage{
musoGlue(preservedCalib, w)
}
\arguments{
\item{plotName}{u}
}
\description{
This function calculates the -users specified- likelihood for random model input.
}
\author{
Roland HOLLOS
}
