% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{musoLeapYears}
\alias{musoLeapYears}
\title{Musoleapyear}
\usage{
musoLeapYears(settings)
}
\description{
How many days are from the given date and given period length(periodlen)?
}
\author{
Roland Hollos
}
\keyword{internal}
