% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMapping.R
\name{musoMapping}
\alias{musoMapping}
\title{musoMapping}
\usage{
musoMapping(code, mapData=NULL)
}
\arguments{
\item{code}{the MuSo outputcode}

\item{mapData}{updateMusomapping generated matrix}
}
\value{
The name of the MuSo outputcode
}
\description{
musoMapping can give us the name of a muso outputcode
}
\author{
Roland Hollos
}
