% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMapping.R
\name{musoMappingFind}
\alias{musoMappingFind}
\title{musoMappingFind}
\usage{
musoMapping(code, mapData=NULL)
}
\arguments{
\item{variable}{If null return the whole mapping, else search a variable code}
}
\value{
The code of th specific name
}
\description{
musoMapping can give us the name of a muso outputcode
}
\author{
Roland Hollos
}
