% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateMuso.R
\name{musoOptimCalib}
\alias{musoOptimCalib}
\title{musoOptimCalib}
\usage{
musoOptimCalib(
  outputLoc = "./",
  calibList = "calibResults.csv",
  parameters = "parameters_soil.csv",
  numTrees = 1000,
  mtry = 3,
  dataTrain = 0.8,
  method = "DEoptim",
  maxIterations = 50,
  NP = 100
)
}
\arguments{
\item{calibList}{The calibration results created by the calibrationMuso, or a path to the csv file with the results.}

\item{parameters}{THe parameters.csv used for setting the min and max values}

\item{numTrees}{Number of trees for the random forest}

\item{mtry}{The number of variables randomly sampled as candidates at each split.}

\item{dataTrain}{The percent of the data to be randomly selected for training. 
Base is 0.8 so 80 percent of the data will be used for training.}

\item{method}{The method to be used for optimization. Currently only DEoptim is available.}

\item{maxIterations}{The maximum number of iterations for the optimization.}

\item{NP}{The population size for the DEoptim algorithm.}
}
\description{
This function allows further post process for calibrateMuso results based on random forest surrogate model which will then be optimized.
Currenlty only DE (differential evolution) is available but in the future custom optimizing method could be used as an input
Note: For multiobjective calibration it won't work, its calibResult file looks a little different but will be supported in the future
}
