% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuso.R
\name{musoParseAndGroupDates}
\alias{musoParseAndGroupDates}
\title{Date parse and grouping helper function}
\usage{
musoParseAndGroupDates(years_to_plot, all_available_dates, silent = TRUE)
}
\description{
Date parse and grouping helper function
}
