% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoRand.R
\name{musoRand}
\alias{musoRand}
\title{musoRand}
\usage{
musoRand(parameters, constrains = NULL, iterations = 3000)
}
\arguments{
\item{parameters}{This is a dataframe (heterogen data-matrix), which first column is the name of the parameters, the second is a numeric vector of the rownumbers of the given variable in the input-file, the last two column consist the endpont of the parameter-ranges, where the parameters will be randomized.}

\item{constrains}{This is a matrics wich specify the constrain rules for the sampling. Further informations coming son.}

\item{iteration}{The number of sample-s. It is adviced to use at least 3000 iteration, because it is generally fast and it can be subsampled later at any time.}
}
\description{
This funtion samples uniformly from the choosen parameters of the BiomeBGC-Muso model, which parameters are constrained by the model logic.
}
\author{
Roland Hollos
}
