% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{musoRandomizer}
\alias{musoRandomizer}
\title{musoRandomizer}
\usage{
musoRandomizer(A, B)
}
\arguments{
\item{A}{is the big matrix}

\item{B}{is the small matrix}
}
\description{
This function is fastly randomize values based on min and max values,
and row indexes.
}
