% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUsefullFunctions.R
\name{musoRedistSoc}
\alias{musoRedistSoc}
\title{Redistribute Soil Organic Carbon (SOC) in a Biome-BGCMuSo Endpoint File}
\usage{
musoRedistSoc(
  input_file,
  output_dir,
  site_name,
  method_name,
  target_layer_cm,
  target_tsoc_layer,
  target_rsoc_layer,
  ratio_components
)
}
\arguments{
\item{input_file}{Path to the original binary endpoint file.}

\item{output_dir}{Path to the directory where the output file will be saved.}

\item{site_name}{Character string for the site name ("askov"). Used for the output filename.}

\item{method_name}{Character string for the method name ("Dist1"). Used for the output filename.}

\item{target_layer_cm}{Numeric. The thickness of the topsoil layer to consider (in cm). Can be any value up to 1000.}

\item{target_tsoc_layer}{Numeric. The target total SOC for the specified topsoil layer.}

\item{target_rsoc_layer}{Numeric. The target resistant SOC (pool 4) for the specified topsoil layer.}

\item{ratio_components}{Numeric vector of length 2 (e.g., c(0.5, 9.5)). Represents the
numerator and denominator of the desired SOC pool 2 to pool 3 ratio. This is needed for distributing the remaning SOC after we've set pool 4. Pool 1 remains untouched.}
}
\value{
Invisibly returns the path to the created output file.
}
\description{
This function reads a binary endpoint file, modifies the SOC (carbon) and SON (nitrogen) pools
based on target values for any specified topsoil depth, and writes a new
endpoint file with a dynamically generated name. The C:N ratio for each
pool is conserved. This function is useful for creating a specific endpoint (initialization) with
our desired SOC distribution for example for a bare fallow experiment with measurements.
}
