% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoSensi.R
\name{musoSensi}
\alias{musoSensi}
\title{musoSensi}
\usage{
musoSensi(
  monteCarloFile = NULL,
  parameters = NULL,
  settings = NULL,
  inputDir = "./",
  outLoc = "./calib",
  outVars = NULL,
  iterations = 30,
  preTag = "mont-",
  outputType = "moreCsv",
  fun = mean,
  varIndex = 1,
  outputFile = "sensitivity.csv",
  plotName = "sensitivity.png",
  plotTitle = "Sensitivity",
  skipSpinup = TRUE,
  skipZero = TRUE,
  sourceFile = NULL,
  postProcString = NULL,
  constraints = NULL,
  modifyOut = TRUE,
  dpi = 300
)
}
\arguments{
\item{parameters}{This is a dataframe (heterogen data-matrix), which first column is the name of the parameters, the second is a numeric vector of the rownumbers of the given variable in the epc-fie, the last two column consist the endpont of the parameter-ranges, where the parameters will be randomized.}

\item{settings}{A list of montecarlos environmental variables. It is generated by the setupMuso() function. In default the settings parameter is generated automatically.}

\item{inputDir}{The location of the input directory, this directory must content a viable pack of all inputfiles and the executable file.}

\item{iterations}{number of the monteCarlo run.}

\item{preTag}{It will be the name of the output files. For example preTag-1.csv, pretag-2csv...}

\item{outputType}{This parameter can be "oneCsv", "moreCsv", and "netCDF". If "oneCsv" is choosen the function create 1 big csv file for all of the runs, if "moreCsv" is choosen, every modell output goes to separate files, if netCDF is selected the outputs will be put in a netCDF file. The default value of the outputTypes is "moreCsv". netCDF is not implemented yet.}

\item{fun}{If you select a variable from the possible outputs (with specify the varIndex parameter), you have to provide a function which maps to a subset of real numbers. The most frequent possibilities are: mean, min, max, var, but you can define any function for your need.}

\item{varIndex}{This parameter specify which parameter of the output will be used. You can extract this information from the ini-files. At the output parameter specifications, the parameters order will determine this number. For example, if you have set these output parameters: 412, 874, 926, 888, and you want to use 926, you should address varIndex with 3.}

\item{skipSpinup}{With this parameter, you can turn of the spinup phase after the first spinup. It will decrease the time significantly.}

\item{calibrationPar}{You may want to change some parameters in your epc file, before you run the modell. You have to select the appropirate modell parameters. You can refence to these with the number of the line in the epc file where the variables are. It indexes from one. You should use a vector for this, like: c(1,5,8)}
}
\description{
This function does regression based sensitivity analysis based on the output of musoMonte.
}
