% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamplog.R
\name{numcut}
\alias{numcut}
\title{This function returns only the starting numbers of a string}
\usage{
numcut(string)
}
\description{
This function returns only the starting numbers of a string
}
\author{
Roland Hollos
}
\keyword{internal}
