% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtableMaker.R
\name{OtableMaker}
\alias{OtableMaker}
\title{OtableMaker}
\usage{
OtableMaker(parametersReal)
}
\arguments{
\item{paramsReal}{The matrix of the parameters}
}
\value{
List of two matrices: A and B matrix for musoRandomizer
}
\description{
This function is generating A and B table for musoRandomizer
}
