% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersweep.R
\name{paramSweep}
\alias{paramSweep}
\title{paramSweep}
\usage{
paramSweep(
  inputDir = "./",
  parameters = NULL,
  outputDir = NULL,
  iterations = 10,
  outVar = 3009,
  fileToChange = "epc",
  htmlOutName = "paramsweep.html",
  fixAlloc = FALSE
)
}
\arguments{
\item{inputDir}{The directory which contains the MuSo model's ini files}

\item{parameters}{A csv file's path which contains the input parameters. The first row must be the name of the parameters, the second is the index of the parameters(row index in the input file), the third is the minimum value of the parameters, the forth is the maximum value of the parameters. If it is not privided, a filebrowser will pop up.}

\item{outputDir}{The path of the directory where the html file will be generated.}

\item{iterations}{The number of changes in the parameter}

\item{outVar}{The name of the output variable to plot, of the MuSo code of it.}

\item{htmlOutName}{The name of the rendered html file}
}
\description{
This function is for testing the modell response to change a set of input variables. It generates an html file which contains a set of graphics of the ...
}
\author{
Roland Hollos
}
