% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuso.R
\name{plotMuso}
\alias{plotMuso}
\title{plot the BBGCMuso output}
\usage{
rungetMuso(settings, timee="d", debugging=FALSE, logfilename=NULL,
keepEpc=FALSE, export=FALSE, silent=FALSE, aggressive=FALSE, leapYear=FALSE)
}
\arguments{
\item{settings}{You have to run the setupMuso function before rungetMuso. It is its output which contains all of the necessary system variables. It sets the whole environment}

\item{variable}{column number of the variable which should be plottedor "all" if you have less than 10 variables. In this case it will plot everything in a matrix layout}

\item{timee}{The required timesteps in the modell output. It can be "d", if it is daily, "m", if it's monthly, "y", it it is yearly}

\item{silent}{If you set it TRUE all off the modells output to the screen will be suppressed. It can be usefull, because it increases the model-speed.}

\item{debugging}{If it is TRUE, it copies the log file to a Log directory to store it, if it is stamplog it contatenate a number before the logfile, which is one more than the maximum of the represented ones in the LOG directory. If it is true or stamplog it collects the "wrong" logfiles}

\item{keepEpc}{If TRUE, it keeps the epc file and stamp it, after these copies it to the EPCS directory. If debugging True or false, it copies the wrong epc files to the wrong epc directory.}

\item{aggressive}{It deletes every possible modell-outputs from the previous modell runs.}

\item{leapyear}{Should the function do a leapyear correction on the outputdata? If TRUE, then the 31.12 day will be doubled.}

\item{export}{if it is yes or you give a filename here, it converts the output to the specific extension. For example, if you set export to "example.csv", it converts the output to "csv", if you set it to "example.xls" it converts to example.xls with the xlsx package. If it is not installed it gives back a warning message and converts it to csv.}
}
\value{
It depends on the export parameter. The function returns with a matrix with the modell output, or writes this in a file, which is given previously
}
\description{
This function runs the BBGC-MuSo model and reads in its outputfile in a very structured way, and after that plot the results automaticly
}
\author{
Roland Hollós
}
