% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuso.R
\name{plotMusoWithData}
\alias{plotMusoWithData}
\title{plot the BBGCMuso output with data}
\usage{
plotMuso(settings, variable,
timee="d", silent=TRUE,
debugging=FALSE, keepEpc=FALSE,
logfilename=NULL, aggressive=FALSE,
leapYear=FALSE, export=FALSE)
}
\arguments{
\item{csvFile}{The file of the measurement. It must contain a header.}

\item{variable}{The name of the output variable to plot}

\item{NACHAR}{This is not implemented yet}

\item{settings}{You have to run the setupMuso function before rungetMuso. It is its output which contains all of the necessary system variables. It sets the whole environment}

\item{sep}{This is the separator used in the measurement file}

\item{savePlot}{It it is specified, the plot will be saved in a format specified with the immanent extension}

\item{calibrationPar}{documentation in setupMuso()}

\item{parameters}{documentation in calibMuso()}
}
\description{
This function runs the BBGC-MuSo model and reads in its outputfile in a very structured way, and after that plot the results automaticly along with a given measurement
}
\author{
Roland Hollos, Dora Hidy
}
