% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcString.R
\name{postProcMuso}
\alias{postProcMuso}
\title{postProcMuso}
\usage{
postProcMuso(modelData, procString, prettyOut = FALSE)
}
\description{
This is a function wich provides some minimal post processing capabilities
}
\keyword{internal}
