% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcpExtremes.R
\name{prcpExtremes}
\alias{prcpExtremes}
\title{prcpExtremes}
\usage{
prcpExtremes(metFileName, skip = 4)
}
\arguments{
\item{metFileName}{The name of the meteorology file (mtc43)}

\item{skip}{number of header lines in meteorology file}
}
\description{
This function calculates the yearly extremes, and averages for a given meteorology file (mtc43)
}
\author{
Erzsebet Kristof
}
