% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeMuso.R
\name{prettyChangeMuso}
\alias{prettyChangeMuso}
\title{prettyChangeMuso}
\usage{
changeMuso(settings, parameters, calibrationPar, fileToChange, fixAlloc)
}
\arguments{
\item{parameters}{the values for the chosen parameters listed in calibrationPar}

\item{calibrationPar}{the line number in the chosen file of the parameters to change}

\item{fileToChange}{defining which file to change}
}
\description{
A function to change lines mainly in soil, epc or ini files
}
