% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistantFunctions.R
\name{readMeasuredMuso}
\alias{readMeasuredMuso}
\title{readMeasuredMuso}
\usage{
readMeasuredMuso(inFile, naString = getOption("datatable.na.strings",
  "NA"), sep = ",", leapYearHandling = TRUE, convert.var = NULL,
  convert.scalar = 1, convert.fun = (function(x) {     x *
  convert.scalar }), convert.file = NULL, filterCol = NULL,
  filterVal = 1, selVar = NULL)
}
\description{
MuSo data reader
}
