% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistantFunctions.R
\name{readObservedData}
\alias{readObservedData}
\title{readMeasuredMuso}
\usage{
readObservedData(
  inFile,
  naString = NULL,
  sep = ",",
  leapYearHandling = TRUE,
  convert.var = NULL,
  convert.scalar = 1,
  convert.fun = (function(x) {     x * convert.scalar }),
  convert.file = NULL,
  filterCol = NULL,
  filterVal = 1,
  selVar = NULL
)
}
\description{
MuSo data reader
}
