% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistantFunctions.R
\name{readValuesFromFile}
\alias{readValuesFromFile}
\title{readValuesFromFile}
\usage{
readValuesFromFile(filename, linums)
}
\arguments{
\item{filename}{The name of the}
}
\description{
read Muso values from file
}
