% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/muso.R
\name{runMuso}
\alias{runMuso}
\title{This runs the Muso model}
\usage{
The function works only, if ...
}
\arguments{
\item{filename}{Name of the initialisation files}
}
\value{
No return, outputs are written to file
}
\description{
This runs the Muso model
}
\author{
Roland
}

