% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.r
\name{setup}
\alias{setup}
\title{This runs the Muso model}
\usage{
setup(executable = NULL, parallel = F, calibrationpar = c(1),
  outputloc = NULL, inputloc = NULL, metinput = NULL, ininput = NULL,
  epcinput = NULL)
}
\arguments{
\item{calibrationpar}{vector with line numbers}
}
\value{
No return, outputs are written to file
}
\description{
This runs the Muso model
}
\author{
Roland
}

