% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupMuso.R
\name{setupMuso}
\alias{setupMuso}
\title{This runs the BBGC-MuSo model}
\usage{
setupMuso(executable = NULL, parallel = F, calibrationpar = c(1),
  outputloc = NULL, inputloc = NULL, metinput = NULL, CO2input = NULL,
  plantinput = NULL, thininput = NULL, mowinput = NULL,
  grazinput = NULL, harvinput = NULL, plouginput = NULL,
  fertinput = NULL, irrinput = NULL, nitinput = NULL, ininput = NULL,
  epcinput = NULL)
}
\arguments{
\item{calibrationpar}{vector with line numbers}
}
\value{
No return, outputs are written to file
}
\description{
This runs the BBGC-MuSo model
}
\author{
Roland Hollós
}
