% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinupMuso.R
\name{spinupMuso}
\alias{spinupMuso}
\title{Run the BBGCMuso modell only in spinup phase, and debugging.}
\usage{
spinupMuso(settings, parameters=NULL, debugging=FALSE,
logfilename=NULL, keepEpc=FALSE, silent=FALSE, aggressive=FALSE)
}
\arguments{
\item{settings}{You have to run the setupMuso function before spinupMuso. It is its output which contains all of the necessary system variables. It sets the whole running environment.}

\item{parameters}{In the settings variable you have set the row indexes of the variables, you wish to change. In this parameter you can give an exact value for them in a vector like: c(1,2,3,4)}

\item{debugging}{If it is TRUE, it copies the log file to a Log directory to store it, if it is stamplog it contatenate a number before the logfile, which is one more than the maximum of the represented ones in the LOG directory. If it is true or stamplog it collects the "wrong" logfiles}

\item{logfilename}{If you want to set a specific name for your logfiles you can set this via logfile parameter}

\item{keepEpc}{If TRUE, it keeps the epc file and stamp it, after these copies it to the EPCS directory. If debugging True or false, it copies the wrong epc files to the wrong epc directory.}

\item{silent}{If you set it TRUE all off the modells output to the screen will be suppressed. It can be usefull, because it increases the model-speed.}

\item{aggressive}{It deletes every possible modell-outputs from the previous modell runs.}
}
\value{
No return, outputs are written to file
}
\description{
This function runs the BBGC-MuSo model's in the spinup phase.
}
\author{
Roland Hollos
}
