% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamplog.R
\name{stamp}
\alias{stamp}
\title{It gives back a stamp wich is the maximum number of the output numcall}
\usage{
stamp(path = "./")
}
\description{
It gives back a stamp wich is the maximum number of the output numcall
}
\author{
Roland Hollos
}
\keyword{internal}
