% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistantFunctions.R
\name{stampAndDir}
\alias{stampAndDir}
\title{stampAndCopy}
\usage{
stampAndDir(outputLoc, names, stampDir, wrongDir, type = "output", errorsign,
  logfiles)
}
\arguments{
\item{outputLoc}{This is the location of the output files.}

\item{outputNames}{These are the prefixis of the logfiles}
}
\value{
Output files with their paths
}
\description{
This function gives us the muso output files with their paths
}
\keyword{internal}
