% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{sumDaysOfPeriod}
\alias{sumDaysOfPeriod}
\title{How many days are from the given date and given period length(periodlen)?}
\usage{
sumDaysOfPeriod(year, periodlen, corrigated = TRUE)
}
\description{
How many days are from the given date and given period length(periodlen)?
}
\author{
Roland Hollos
}
\keyword{internal}
