% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUsefullFunctions.R
\name{supportedMuso}
\alias{supportedMuso}
\title{supportedMuso}
\usage{
supportedMuso(type="outputs")
}
\arguments{
\item{type}{"outputs" or "message", if you choose "outputs", it gives you a simple vector of the formats, if you choose "message", it gives you a full sentence which contains the same information.}
}
\value{
if you choose "outputs", it gives you a simple vector of the formats, if you choose "message", it gives you a full sentence which contains the same information.
}
\description{
A function for getting the list of the output formats which is supported by RBBGCMuso
}
\author{
Roland Hollós
}
