% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunerTest.R
\name{tuneMuso2}
\alias{tuneMuso2}
\title{tuneMusoTest}
\usage{
tuneMuso2(directory = NULL, ...)
}
\arguments{
\item{...}{Other parameters for shinyApp function}
}
\description{
launchApp launch the shiny app
}
