% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunerTest.R
\name{tuneMusoServerTest}
\alias{tuneMusoServerTest}
\title{tuneMusoServer}
\usage{
...
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{dinamic session management for shiny}
}
\description{
Server program for tumeMuso
}
