% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunerTest.R
\name{tuneMusoTest}
\alias{tuneMusoTest}
\title{tuneMusoTest}
\usage{
tuneMusoTest(directory = NULL, ...)
}
\arguments{
\item{...}{Other parameters for shinyApp function}
}
\description{
launchApp launch the shiny app
}
