% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\name{tuneMusoUI}
\alias{tuneMusoUI}
\title{tuneMusoUI}
\usage{
...
}
\arguments{
\item{parameterFile}{optional, the parameter csv file}
}
\description{
This is a simple parameter tuner function which works great in a flat directory systemj
}
