% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMapping.R
\name{updateMusoMapping}
\alias{updateMusoMapping}
\title{updateMusoMapping}
\usage{
updateMusoMapping(output_map_init="output_map_init.c")
}
\arguments{
\item{output_map_init}{The outputcode-variable bindings is described in output_map_init.c file, via this parameter, you can give it the the function}
}
\value{
The outputcode-variable matrix, and also change the global variable
}
\description{
This function update the the muso outputcode-variable matrix
}
\author{
Roland Hollos
}
