% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{alignData}
\alias{alignData}
\title{alignData}
\usage{
alignData(mdata, dataCol, modellSettings = NULL, startDate = NULL,
  endDate = NULL, formatString = "\%Y-\%m-\%d", leapYear = TRUE,
  continious = TRUE)
}
\description{
This function align the data to the model and the model to the data
}
\keyword{internal}
