% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeMuso.R
\name{changemulline}
\alias{changemulline}
\title{changemulline}
\usage{
changemulline(
  filePaths,
  calibrationPar,
  contents,
  src = NULL,
  outFiles = filePaths
)
}
\description{
The function uses the previous changspecline function to operate.
}
\author{
Roland Hollos
}
