% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeMuso.R
\name{changemulline}
\alias{changemulline}
\title{This is the function which is capable change multiple specific lines to other using their row numbers.}
\usage{
changemulline(filename, calibrationpar, contents)
}
\description{
he function uses the previous changspecline function to operate.
}
\author{
Roland Hollos
}
\keyword{internal}
