% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanupMuso}
\alias{cleanupMuso}
\title{cleanupMuso}
\usage{
cleanupMuso(location=NULL, simplicity=TRUE,deep=FALSE)
}
\arguments{
\item{location}{This is the place (directory) where your output files are located.}

\item{simplicity}{TRUE or FALSE. If TRUE cleanupMuso will erase only the log files from the location}

\item{deep}{If it is TRUE, it will delete every files from the subdirectories also}
}
\description{
cleanupMuso can erase all of the unnecessary log and output files.
}
\author{
Roland HOLLOS
}
