% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuso.R
\name{compareMuso}
\alias{compareMuso}
\title{compareMuso}
\usage{
compareMuso(settings, parameters, variable, calibrationPar = NULL,
  fileToChange = NULL, skipSpinup = TRUE, timeFrame = "day")
}
\arguments{
\item{settings}{You have to run the setupMuso function before rungetMuso. It is its output which contains all of the necessary system variables. It sets the whole environment}

\item{parameters}{In the settings variable you have set the row indexes of the variables, you wish to change. In this parameter you can give an exact value for them in a vector like: c(1,2,3,4)}

\item{variable}{The name of the output variable to plot}

\item{calibrationPar}{in the help of setupMuso function.}

\item{fileToChange}{You can change any line of the epc or the ini file, you just have to specify with this variable which file you van a change. Two options possible: "epc", "ini", "both"}
}
\description{
This function runs the modell, change one of it's input, and plot both in one plot.
}
\author{
Roland Hollos
}
