% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoExamples.R
\name{copyMusoExamleTo}
\alias{copyMusoExamleTo}
\title{copyMusoExampleTo}
\usage{
copyMusoExamleTo(example = NULL, destination = NULL)
}
\arguments{
\item{example}{The name of the example file, if it is NULL tcl/tk menu will pop up to select.}
}
\value{
List of two matrices: A and B matrix for musoRandomizer
}
\description{
With this function you can copy RBBGCMuso example library whereever you want
}
