% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilQuery.R
\name{createSoilFile}
\alias{createSoilFile}
\title{createSoilFile}
\usage{
createSoilFile(
  lat,
  lon,
  outputFile = "recent.soi",
  method = "constant",
  apiURL,
  getVWC = FALSE,
  template = system.file("examples/hhs/hhs_muso7.soi", package = "RBBGCMuso")
)
}
\description{
This function collects soil data from a given restapi, de default is soilGrid
}
\author{
Roland HOLLOS
}
