% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/other_usefull_functions.R
\name{getyearlycum}
\alias{getyearlycum}
\title{'Funtion for getting cumulative yearly data from observations}
\usage{
getyearlycum(daily_observations)
}
\arguments{
\item{A}{vector of the daily observations.}
}
\value{
A vector of yearly data
}
\description{
'Funtion for getting cumulative yearly data from observations
}
\author{
Roland Hollós
}

