% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metReportMuso.R
\name{metReportMuso}
\alias{metReportMuso}
\title{metReportMuso}
\usage{
metReportMuso(metFileName, skip = 4, numericReport = FALSE)
}
\arguments{
\item{metFileName}{The name of the meteorology file (mtc43)}

\item{skip}{number of header lines in meteorology file}
}
\description{
This function calculates the daily and yearly statistics for a given meteorology file (mtc43).
Statistics are available as text and numeric report.
}
\author{
Erzsebet Kristof
}
