% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantManagement.R
\name{mgmTimePlot}
\alias{mgmTimePlot}
\title{mgmTimePlot}
\usage{
mgmTimePlot(mgmDir = "./")
}
\arguments{
\item{mgmDir}{This folder contains the management files.}
}
\value{
An interactive plot about time and management.
}
\description{
This function shows the timeline of the plant management types.
}
\author{
Erzsebet Kristof
}
