% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUsefullFunctions.R
\name{musoBatchRedistSoc}
\alias{musoBatchRedistSoc}
\title{Batch Process SOC Redistribution for Multiple Scenarios}
\usage{
musoBatchRedistSoc(config)
}
\arguments{
\item{config}{A list containing all parameters for the batch run. See example for structure}
}
\value{
Invisibly returns a vector of the file paths created.
}
\description{
A wrapper function to run `musoRedistSoc` for multiple methods and ratios
based on a single configuration list.
}
\examples{
run_config <- list(
 
 # Site level information
 site_name = "site1",
 output_dir = "c:/muso/SOC/",
 # this will be the base endpoint file which will be modified for each method and ratio
 input_file = "c:/muso/SOC/base.endpoint", 
 
 target_layer_cm = 23,
 target_tsoc_layer = 4.77,
 
 # "method" names and their corresponding rsoc values
 methods = list(
   list(name = "Method1", rsoc = 1.32),
   list(name = "Method2", rsoc = 0.4),
   list(name = "Method3", rsoc = 1.76)
 ),
 
 # A list of ratio components. The script will loop through these for EACH method
 ratios = list(
   c(1, 9), c(2, 8), c(3, 7), c(4, 6), c(5, 5),
   c(6, 4), c(7, 3), c(8, 2), c(9, 1),
   c(0.5, 9.5) 
 )
)
# using the function with the example list
musoBatchRedistSoc(config = run_config)

}
