% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{musoDate}
\alias{musoDate}
\title{It generates BiomeBGC-MuSo dates}
\usage{
musoDate(settings, timestep = "d", combined = TRUE, corrigated = TRUE,
  format = "en")
}
\arguments{
\item{settings}{You have to run the setupMuso function before musoDate. It is its output which contains all of the necessary system variables. It sets the whole environment.}

\item{timestep}{timestep, which can be daily ("d"), monthly ("m"), yearly("y")}

\item{combined}{If FALSE the output is a vector of 3 string: day, month year, if true, these strings will be concatenated.}

\item{corrigated}{If True it counts with leapyears, else dont.}

\item{format}{This is the format of the date. It can be "en" (dd.mm.yyyy), or "hu" (yyyy.mm.dd)}
}
\value{
The exact date-vectors for the BioBGC-MuSo output. You can use this for labelling purpose for example.
}
\description{
It generates all of the day-dates which are between the start and endyear of BiomeBGC-MuSo run.
How many days are from the given date and given period length(periodlen)?
}
\author{
Roland Hollos
}
