% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoTime.R
\name{musoDate}
\alias{musoDate}
\title{musoDate}
\usage{
musoDate(
  startYear,
  endYears = NULL,
  numYears,
  combined = TRUE,
  leapYearHandling = FALSE,
  prettyOut = FALSE
)
}
\arguments{
\item{startYear}{Start year of the simulations}

\item{numYears}{Number of the years of the simulation}

\item{combined}{using separate y m d columns or not?}

\item{timestep}{timestep of date creation}

\item{corrigated}{If leapyear ...}

\item{format}{"the date format"}
}
\description{
This function generates MuSo compatibla dates for the data
}
\author{
Roland HOLLOS
}
