% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuso.R
\name{musoEnsemblePlot}
\alias{musoEnsemblePlot}
\title{Create Ensemble Plot}
\usage{
musoEnsemblePlot(
  run_output_subfolder = "thread",
  measurement_data,
  model_mapping_code,
  measurement_data_column,
  plot_individual_lines = TRUE,
  year_axis_interval = 2,
  best_run_param_file = "maxlikelihood_parameters.csv",
  fileToChange = "epc",
  settings = setupMuso(),
  output_plot_filename_prefix = "ensemble_plot"
)
}
\arguments{
\item{run_output_subfolder}{Character. Name of the subfolder within `working_directory`
that contains the individual run output folders by default calibrateMuso creates a "thread" folder in the root dir which in turn
contains "thread_1/calib/", "thread_2/calib/", etc.}

\item{measurement_data}{R object or Character. Either a data.frame/data.table object
containing the measurement data, or a character string path to the measurement CSV file.}

\item{model_mapping_code}{Numeric. The model variable code used by
`musoMapping()` to get the variable name.}

\item{measurement_data_column}{Character. The name of the column in
`measurement_data` to be used for plotting observed values.}

\item{plot_individual_lines}{Logical. If `TRUE`, plots individual ensemble member lines.
If `FALSE`, plots ensemble summary (median and quantiles). Default is `TRUE`.}

\item{year_axis_interval}{Numeric. Interval for year ticks on the x-axis. Default is `2`.}

\item{best_run_param_file}{Character. Path to the CSV file containing parameters for the
"best run" simulation (most likely "maxlikelihood_parameters.csv"). This path should be
relative to `working_directory` or an absolute path. If `NULL` or file not found,
the best run line is not plotted. Default is "maxlikelihood_parameters.csv" since calibrateMuso creates this by default.}

\item{fileToChange}{Character. The type of MuSo file to change for the best run
simulation, typically "epc" or "soil". Default is "epc".}

\item{settings}{An RBBGCMuso settings object (output of `setupMuso()`).
Required if plotting the best run. If `NULL` and best run is attempted,
a warning will be issued.}

\item{output_plot_filename_prefix}{Character. Prefix for the output PNG filename.
The plot type (individual/summary) will be appended. Default "ensemble_plot".}
}
\value{
Invisibly returns the ggplot object. Saves the plot to a PNG file in `working_directory`.
}
\description{
Generates an ensemble plot from model run outputs, including
measurement data and a "best run" simulation.
}
\details{
The function expects model run output CSVs to be structured as:
`working_directory/run_output_subfolder/thread_*/calib/*.csv`.
The first column of these CSVs should be dates parsable by `as.Date` with format "%d.%m.%Y".
The `best_run_param_file` is expected to have a structure that `changeMuso`
can use, typically with parameter values in the third column and calibration parameter names/codes
in the second. This one is provided by calibrateMuso

Currently this function uses one binded data.table for all ensemble members so it can use high amounts of memory.

Optionally you can use the 'ragg' package for a faster plotting, but it is not required.
}
