% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musoMonte.R
\name{musoMonte}
\alias{musoMonte}
\title{musoMonte}
\usage{
musoMonte(settings = NULL, parameters, inputDir = "./",
  outLoc = "./calib", iterations = 10, preTag = "mont-",
  inputName = paste0(pretag, "epcs.csv"), outputType = "moreCsv",
  fun = mean, varIndex = 8, doSensitivity = FALSE, onDisk = FALSE, ...)
}
\arguments{
\item{settings}{Do you want to run multiple modell paralelly, if yes, set this variable to TRUE}

\item{parameters}{stores the place of the modell-executable file. In normal usage, you don't have to be set this, because a RBBgcmuso package contains allways the latest modell executable. In spite of this, if you would like to use this package for modell development or just want to use different models (for example for comparison), you will find it useful}

\item{outLoc}{Where should the modell puts its outputs. You should give a location for it via this variable, for example: outputLoc="/place/of/the/outputs/"}

\item{iterations}{Usually it is the root directory, where you put the iniFiles for the modell}

\item{preTag}{Via metInput parameter, you can tell the modell where are the meteorological files. As default it reads this from the iniFiles.}

\item{inputName}{Via CO2 parameter, you can tell the modell where are the CO2 data files. As default it reads this from the iniFiles.}

\item{outputType}{Via planting parameter, you can tell the modell where are the data files, which contains the planting informations. As default it reads this from the iniFiles.}

\item{fun}{Via thining parameter, you can tell the modell where are the data files, which contains the thining informations. As default it reads this from the iniFiles.}

\item{varIndex}{Via mowing parameter, you can tell the modell where are the data files, which contains the mowing informations. As default it reads this from the iniFiles.}

\item{doSensitivity}{Via grazing parameter, you can tell the modell where are the data files, which contains the grazing informations. As default it reads this from the iniFiles.}

\item{onDisk}{Via harvesting parameter, you can tell the modell where are the data files, which contains the harvesting informations. As default it reads this from the iniFiles.}

\item{calibrationPar}{You may want to change some parameters in your epc file, before you run the modell. You have to select the appropirate modell parameters. You can refence to these with the number of the line in the epc file where the variables are. It indexes from one. You should use a vector for this, like: c(1,5,8)}
}
\description{
This funcion is fundamental for the BiomBGC-MuSo modell related functions like spinupMuso, normalMuso, rungetMuso, because it sets the modells environment.
inputDir = "./",
}
\author{
Roland Hollos
}
