% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateMuso.R
\name{musoOptimCalib}
\alias{musoOptimCalib}
\title{musoOptimCalib}
\usage{
musoOptimCalib(
  outputLoc = "./",
  calibList = "calibResults.csv",
  parameters = "parameters_soil.csv",
  numTrees = 1000,
  mtry = 3,
  dataTrain = 0.8,
  method = "DEoptim",
  maxIterations = 50,
  NP = 100,
  saveAllNP = FALSE,
  parallel = FALSE,
  cluster = NULL,
  numCores = parallel::detectCores() - 1
)
}
\arguments{
\item{calibList}{The calibration results created by the calibrationMuso, or a path to the csv file with the results.}

\item{parameters}{THe parameters.csv used for setting the min and max values}

\item{numTrees}{Number of trees for the random forest}

\item{mtry}{The number of variables randomly sampled as candidates at each split.}

\item{dataTrain}{The percent of the data to be randomly selected for training. 
Base is 0.8 so 80 percent of the data will be used for training.}

\item{method}{The method to be used for optimization. Currently only DEoptim is available.}

\item{maxIterations}{The maximum number of iterations for the optimization.}

\item{NP}{The population size for the DEoptim algorithm. Generally setting this 10 times larger than your parameter vector is necessary for proper results}

\item{saveAllNP}{saves all the population results as the calibration runs and writes it into a csv later}

\item{parallel}{option to perform the optimization method in a parallel way for better speed}

\item{cluster}{for better perfomance in case you call this function in a loop this option allows the use of already created clusters. If given the musoOptimCalib function won't create and destroy clusters at each call.}

\item{numCores}{Number of cores to be used if parallel is TRUE}
}
\description{
This function allows further post process for calibrateMuso results based on random forest surrogate model which will then be optimized.
Currenlty only DE (differential evolution) is available but in the future custom optimizing method could be used as an input
Note: For multiobjective calibration it won't work, its calibResult file looks a little different but will be supported in the future
}
