% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickeffect.R
\name{musoQuickEffect}
\alias{musoQuickEffect}
\title{musoQuickEffect}
\usage{
musoQuickEffect(settings = NULL, calibrationPar = NULL, startVal,
  endVal, nSteps = 1, fileToChange = "epc", outVar,
  parName = "parVal")
}
\arguments{
\item{settings}{The settings from setupMuso output}

\item{startVal}{The oroginal parameterValue}

\item{endVal}{The goal value while the function pass}

\item{nSteps}{How many steps 'till you reach the endVal}

\item{fileTochange}{Please choose "epc" "ini" or "both". This is the place of the orininal variable.}
}
\value{
An effect plot
}
\description{
This function changes a choosen parameter, and visualize the effect of the change on a chosen variable.
}
\author{
Roland Hollos
}
