% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickeffect.R
\name{musoQuickEffect}
\alias{musoQuickEffect}
\title{musoQuickEffect}
\usage{
musoQuickEffect(
  settings = setupMuso(),
  calibrationPar = NULL,
  startVal,
  endVal,
  nSteps = 1,
  fileToChange = "epc",
  modifyOriginal = TRUE,
  outVar,
  parName = "parVal",
  yearNum = 1,
  year = (settings$startYear + yearNum - 1)
)
}
\arguments{
\item{settings}{RBBGCMuso uses variables that define the entire simulation environment. Those environment variables include the name of the INI files, the name of the meteorology files, the path to the model executable and its file name, the entire output list, the entire output variable matrix, the dependency rules for the EPC parameters etc. Using the runMuso function RBBGCMuso can automatically create those environment variables by inspecting the files in the working directory (this happens through the setupMuso function). It means that by default model setup is performed automatically in the background and the user has nothing to do. With this settings parameter we can force runMuso to skip automatic environment setup as we provide the environment settings to runMuso. In a typical situation the user can skip this option.}

\item{startVal}{The initial value of the given parameter.}

\item{endVal}{The maximum of the given parameter.}

\item{nSteps}{Number of steps from startVal to endVal. It equals the number of simulations, and number of curves on the final plot.}

\item{fileTochange}{Please choose "EPC", "INI" or "BOTH". This file will be used for the analysis, and the original parameter values will be changed according to the choice of the user.}
}
\value{
Graph showing the runs with the selected parameters with color coding. The graph will show data from the last simulation year.
}
\description{
This function changes a chosen parameter from the INI or from the ecophysiological constants file (EPC) within a predefined range (defined by the user), and visualizes the effect of the change on the selected output variable. The user has to specify the parameter, the interval for the parameter effect test, and the number of steps. This function focuses only on one parameter. The so-called paramSweep function can manipulate multiple INI/EPC parameters and visualize the results.
}
\author{
Roland HOLLOS
}
