% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{optiMuso}
\alias{optiMuso}
\title{optiMuso}
\usage{
optiMuso(measuredData, parameters = NULL, startDate = NULL,
  endDate = NULL, formatString = "\%Y-\%m-\%d",
  leapYearHandling = TRUE, dataVar, outLoc = "./calib",
  preTag = "cal-", settings = NULL, outVars = NULL,
  iterations = 30, skipSpinup = TRUE, constrains = NULL,
  plotName = "calib.jpg", likelihood = function(x, y) {    
  exp(-sqrt(mean((x - y)^2))) }, continious, modelVar = 3009,
  naVal = NULL, postProcString = NULL)
}
\arguments{
\item{parameters}{b}

\item{startDate}{d}

\item{endDate}{e}

\item{formatString}{a}

\item{outLoc}{c}

\item{settings}{e}

\item{iterations}{c}

\item{skipSpinup}{a}

\item{constrains}{d}

\item{plotName}{u}

\item{likelihood}{d}

\item{measuredDataFile}{a}

\item{sep}{c}

\item{filterCol}{a}

\item{filterVal}{b}

\item{selVar}{c}

\item{pretag}{a}

\item{calPar}{a}

\item{leapYear}{b}
}
\description{
This function calculates the -users specified- likelihood for random model input.
}
\author{
Roland HOLLOS
}
