% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{optiMuso}
\alias{optiMuso}
\title{optiMuso}
\usage{
optiMuso(measuredDataFile, parameters = NULL, sep = ",", startDate,
  endDate, formatString = "\%Y-\%m-\%d", naString = NULL,
  leapYear = TRUE, filterCol = NULL, filterVal = 1, selVar,
  outLoc = "./calib", preTag = "cal-", settings = NULL,
  outVars = NULL, iterations = 30, skipSpinup = TRUE,
  constrains = NULL, plotName = "calib.jpg", likelihood = function(x,
  y) {     exp(-sqrt(mean((x - y)^2))) }, calPar = 3009)
}
\arguments{
\item{measuredDataFile}{a}

\item{parameters}{b}

\item{sep}{c}

\item{startDate}{d}

\item{endDate}{e}

\item{formatString}{a}

\item{leapYear}{b}

\item{filterCol}{a}

\item{filterVal}{b}

\item{selVar}{c}

\item{outLoc}{c}

\item{settings}{e}

\item{iterations}{c}

\item{skipSpinup}{a}

\item{constrains}{d}

\item{plotName}{u}

\item{likelihood}{d}

\item{calPar}{a}

\item{pretag}{a}
}
\description{
This function calculates the -users specified- likelihood for random model input.
}
\author{
Roland HOLLOS
}
