% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuso.R
\name{plotMusoWithData}
\alias{plotMusoWithData}
\title{plot the Biome-BGCMuSo model output with observation data}
\usage{
plotMuso(settings, variable,
timee="d", silent=TRUE,
debugging=FALSE, keepEpc=FALSE,
logfilename=NULL, aggressive=FALSE,
leapYear=FALSE, export=FALSE)
}
\arguments{
\item{csvFile}{This specifies the filename of the measurements. It must contain a header. Typically this is a CSV file.}

\item{variable}{The name of the output variable to plot}

\item{NACHAR}{This is not implemented yet}

\item{settings}{RBBGCMuso uses variables that define the entire simulation environment. Those environment variables include the name of the INI files, the name of the meteorology files, the path to the model executable and its file name, the entire output list, the entire output variable matrix, the dependency rules for the EPC parameters etc. Using the runMuso function RBBGCMuso can automatically create those environment variables by inspecting the files in the working directory (this happens through the setupMuso function). It means that by default model setup is performed automatically in the background and the user has nothing to do. With this settings parameter we can force runMuso to skip automatic environment setup as we provide the environment settings to runMuso. In a typical situation the user can skip this option.}

\item{sep}{This is the separator symbol used in the measurement file (that is supposed to be a delimited text file)}

\item{savePlot}{It it is specified, the plot will be saved in a graphical format specified by the immanent extension. For example, it the savePlot is set to image01.png then a PNG graphics file will be created.}

\item{calibrationPar}{You might want to change some parameters in your EPC file before running the model. The function offers possibility for this without editing the EPC file. In this situation you have to select the appropirate model parameters first. You can refer to these parameters with the number of the line in the EPC file. Indexing of lines start from one. You should use a vector for this referencing like c(1,5,8)}

\item{parameters}{Using the function it is possible to change some of the EPC parameters prior to model execution. This can be achieved with this option. In the parameters variable you have set the row indices of the variables that you wish to change. In this parameters you can give an exact value for them in a vector form like c(1,2,3,4).}
}
\description{
This function runs the Biome-BGCMuSo model and reads its output file in a well structured way, and after that it plots the results automatically along with a given measurement dataset provided by the user. plotMusoWithData is a convenient and quick method to create nice graphs from Biome-BGCMuSo output which is quite painful in other environments.
}
\author{
Roland HOLLOS, Dora HIDY
}
