% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinupMuso.R
\name{spinupMuso}
\alias{spinupMuso}
\title{Runs the Biome-BGCMuSo model in spinup phase (execution of normal phase is possible with normalMuso) with debugging features.}
\usage{
spinupMuso(settings, parameters=NULL, debugging=FALSE,
logfilename=NULL, keepEpc=FALSE, silent=FALSE, aggressive=FALSE)
}
\arguments{
\item{settings}{RBBGCMuso uses variables that define the entire simulation environment. Those environment variables include the name of the INI files, the name of the meteorology files, the path to the model executable and its file name, the entire output list, the entire output variable matrix, the dependency rules for the EPC parameters etc. Using the runMuso function RBBGCMuso can automatically create those environment variables by inspecting the files in the working directory (this happens through the setupMuso function). It means that by default model setup is performed automatically in the background and the user has nothing to do. With this settings parameter we can force runMuso to skip automatic environment setup as we provide the environment settings to runMuso. In a typical situation the user can skip this option.}

\item{parameters}{|||| In the parameters variable you have set the row indices of the variables that you wish to change. In this parameter you can provide an exact value for them in a vector form like c(1,2,3,4)}

\item{debugging}{If debugging is set to TRUE, after model execution the function copies the Biome-BGCMuSo log file into a LOG directory to stores it for further processing. If debugging is set to STAMPLOG instead of TRUE, it concatenates a number before the logfile, which is one plus the maximum of those present in the LOG directory. In each case the log files will be saved.}

\item{logfilename}{If you would like to set a specific name for the logfiles you can set this via the logfilename parameter}

\item{keepEpc}{If keepEpc is set to TRUE, the function keeps the EPC file and stamps it, and then copies it to the EPCS directory. If debugging is set to TRUE, it copies the wrong EPC files to the wrong epc directory.}

\item{silent}{If you set the silent parameter to TRUE, all of the model's output normally written to the screen will be suppressed. This option can be useful to increase the speed of the model execution.}

\item{aggressive}{It deletes all previous model-outputs from previous model runs.}
}
\value{
No return, outputs are written to file
}
\description{
This function runs the Biome-BGCMuSo model in spinup phase.
}
\author{
Roland HOLLOS
}
