% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMapping.R
\name{updateMusoMapping}
\alias{updateMusoMapping}
\title{updateMusoMapping}
\usage{
updateMusoMapping(excelName, dest = "./", version = getOption("RMuso_version"))
}
\arguments{
\item{excelName}{Name of the excelfile which contains the parameters}
}
\value{
The output code-variable matrix, and also the function changes the global variable
}
\description{
This function updates the Biome-BGCMuSo output code-variable matrix. Within Biome-BGCMuSo the state variables and fluxes are marked by integer numbers. In order to provide meaningful variable names (e.g. 3009 means Gross Primary Production in Biome-BGCMuSo v5) a conversion table is needed which is handled by this function.
}
\author{
Roland HOLLOS
}
